refactor into smaller components and improve ui
